calibration/requirements.txt
# calibration/requirements.txt
# Install with: pip install -r requirements.txt

numpy>=1.26.0
scipy>=1.12.0
PyYAML>=6.0.1

# Optional — for diagnostic plots:
# matplotlib>=3.8.0
